requirements.txt
tavily
langchain
langchain-community
langchain-openai
langchain-text-splitters
pymupdf
pillow
opencv-python
unstructured
huggingface_hub
google-generativeai
numpy
tiktoken
tqdm
python-dotenv
requests
streamlit
sentencepiece
torch
transformers
lancedb
openai
sentence-transformers
rank_bm25